Stack is Empty Error 해결 요망

+++ app/src/main/java/com/example/mobile_programming_term_project/makeCalculationStack.java
@@ -4,6 +4,6 @@ public interface makeCalculationStack {
     boolean isEmpty();
     boolean isFull();
     void push(int item);
-    double pop();
+    int pop();
     int peek();
 }
